Move saving/loading code to save.cpp
