+++ src/elona/Android.mk
@@ -88,6 +88,7 @@ quest.cpp \
 race.cpp \
 random.cpp \
 random_event.cpp \
+save.cpp \
 set_item_info.cpp \
 shop.cpp \
 spec.cpp \
