+++ src/elona/CMakeLists.txt
@@ -71,6 +71,7 @@ set(ELONA_SOURCES
   race.cpp
   random.cpp
   random_event.cpp
+  save.cpp
   set_item_info.cpp
   shop.cpp
   spec.cpp
