+++ src/elona/save.hpp
@@ -1,0 +1,12 @@
+#pragma once
+
+
+
+namespace elona
+{
+
+void load_save_data();
+void do_save_game();
+void save_game();
+
+} // namespace elona
